requirements.txt
langchain_community
langchain_experimental
langchain-huggingface
langchain-ollama
langchain-pinecone
langchain-postgres
langchain-chroma
langchain-weaviate
langchain-deeplake
langchain-pinecone
faiss-cpu
weaviate-client
deeplake
pinecone-text
python-dotenv
langchain_groq
pandas
streamlit
pymupdf4llm
PyPDF2
selenium
webdriver-manager
pytesseract
camelot-py
tabula-py[jpype]
ollama